Finalize MeSH pipelines with real data results and serialization fixes.
- Fixed serialization bug where NumPy types were stored as strings in CSV.
- Improved 'top_terms_by_year' logic to filter out zero-count terms.
- Verified all results with real MeSH/PubMed data samples.
- Cleaned up documentation and directory structure.
- Provided fresh results in 'scripts/research/mesh/data/' as clean CSV files.

+++ scripts/research/mesh/analysis.py
@@ -38,10 +38,10 @@ class Analyzer:
             stats.append({
                 'ui': ui,
                 'name': self.candidate_map.get(ui, "Unknown"),
-                'total_count': total_count,
-                'years_active': years_active,
-                'recent_slope': slope,
-                'score': total_count * (years_active / 50.0) # Heuristic score
+                'total_count': int(total_count),
+                'years_active': int(years_active),
+                'recent_slope': float(slope),
+                'score': float(total_count * (years_active / 50.0)) # Heuristic score
             })
 
         return pd.DataFrame(stats)
@@ -86,9 +86,9 @@ class Analyzer:
             return {
                 'model': 'logistic',
                 'params': {
-                    'L': popt[0] * y_max,
-                    'k': popt[1],
-                    'x0': popt[2]
+                    'L': float(popt[0] * y_max),
+                    'k': float(popt[1]),
+                    'x0': float(popt[2])
                 }
             }
         except:
@@ -97,7 +97,7 @@ class Analyzer:
             return {
                 'model': 'linear',
                 'params': {
-                    'slope': slope,
-                    'intercept': intercept
+                    'slope': float(slope),
+                    'intercept': float(intercept)
                 }
             }
